Ultima version espero

+++ src/com/example/otp_android/MySQLiteHelper.java
@@ -18,7 +18,7 @@ public class MySQLiteHelper extends SQLiteOpenHelper {
     public static final String COLUMN_COUNTER = "counter";
 
     private static final String DATABASE_NAME = "users.db";
-    private static final int DATABASE_VERSION = 1;
+    private static final int DATABASE_VERSION = 2;
 
     // Database creation sql statement
     private static final String DATABASE_CREATE = "create table "
@@ -40,8 +40,10 @@ public class MySQLiteHelper extends SQLiteOpenHelper {
 
     @Override
     public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
-        db.execSQL("DROP TABLE IF EXISTS " + TABLE_USERS);
-        onCreate(db);
+    	if(newVersion > oldVersion){                    
+    		 db.execSQL("DROP TABLE IF EXISTS " + TABLE_USERS);
+    	     onCreate(db);
+    }
     }
 
 }
